Little refactoring: move splash script to lua file
